// client/src/lib/vendo.dto.ts
/* 
 * Autogenerated by backend build - Do not edit here! 
 * Run "npm run build" from the server directory to update this file. 
 */

export type Station = {
    id: string;
    name: string;
}
export type TimetableEntry = {
    type: 'arrival' | 'departure' | 'stopover';
    line?: string;
    lineType?: string;
    delay?: number;
    platform?: string;
    arrival?: string;
    departure?: string;
    origin?: string;
    destination?: string;
    tripId: string;
}